Restructure header files
