+++ src/enzyme_ad/jax/Dialect/Comm/Comm.h
@@ -1,0 +1,8 @@
+#ifndef ENZYME_AD_JAX_DIALECTS_COMM_COMM_H
+#define ENZYME_AD_JAX_DIALECTS_COMM_COMM_H
+
+#include "src/enzyme_ad/jax/Dialect/Comm/CommDialect.h"
+#include "src/enzyme_ad/jax/Dialect/Comm/CommTypes.h"
+#include "src/enzyme_ad/jax/Dialect/Comm/CommOps.h"
+
+#endif
