+++ src/enzyme_ad/jax/Dialect/Comm/CommDialect.cpp
@@ -1,29 +1,20 @@
-#include "src/enzyme_ad/jax/Dialect/Comm/CommDialect.h"
-#include "src/enzyme_ad/jax/Dialect/Comm/CommDialect.cpp.inc"
+#include "src/enzyme_ad/jax/Dialect/Comm/Comm.h"
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/ADT/TypeSwitch.h"
 
 using namespace mlir;
 using namespace mlir::comm;
 
-void CommunicationDialect::initialize() {
+#include "src/enzyme_ad/jax/Dialect/Comm/CommDialect.cpp.inc"
+
+void CommDialect::initialize() {
   addTypes<
-#define GET_TYPEDEF_LIST
-#include "src/enzyme_ad/jax/Dialect/Comm/CommTypes.cpp.inc"
-      >();
+    #define GET_TYPEDEF_LIST
+    #include "src/enzyme_ad/jax/Dialect/Comm/CommTypes.cpp.inc"
+  >();
 
-  addAttributes<
-#define GET_ATTRDEF_LIST
-#include "src/enzyme_ad/jax/Dialect/Comm/CommAttrs.cpp.inc"
-      >();
   addOperations<
-#define GET_OP_LIST
-#include "src/enzyme_ad/jax/Dialect/Comm/CommOps.cpp.inc"
-      >();
+    #define GET_OP_LIST
+    #include "src/enzyme_ad/jax/Dialect/Comm/CommOps.cpp.inc"
+  >();
 }
-
-#define GET_ATTRDEF_CLASSES
-#include "src/enzyme_ad/jax/Dialect/Comm/CommAttrs.cpp.inc"
-
-#define GET_TYPEDEF_CLASSES
-#include "src/enzyme_ad/jax/Dialect/Comm/CommTypes.cpp.inc"
