+++ src/enzyme_ad/jax/Dialect/Comm/CommInterfaces.cpp
@@ -1,4 +1,4 @@
-#include "src/enzyme_ad/jax/Dialect/Comm/CommDialect.h"
+#include "src/enzyme_ad/jax/Dialect/Comm/CommOps.h"
 
 using namespace mlir;
 using namespace mlir::comm;
