+++ src/enzyme_ad/jax/Passes/Passes.td
@@ -14,14 +14,14 @@ include "mlir/Pass/PassBase.td"
 def CommExplodeSplits : Pass<"comm-explode-splits"> {
   let summary = "Converts a (nested) split into a non-nested split with one branch per device";
   let dependentDialects = [
-    "comm::CommunicationDialect"
+    "comm::CommDialect"
   ];
 }
 
 def CommDeplex : Pass<"comm-deplex"> {
   let summary = "Removes multiplex messages by replacing each use of the output token with one of the input tokens";
   let dependentDialects = [
-    "comm::CommunicationDialect"
+    "comm::CommDialect"
   ];
 }
 
@@ -30,7 +30,7 @@ def CommRemoveDeadMessages : Pass<"comm-remove-dead-messages"> {
   let description = [{This does NOT perform a full analog to DCE- this will leave in useless communications if they
   have any dependent recieve operation, regardless of whether the result of the receive is a dead value.}];
   let dependentDialects = [
-    "comm::CommunicationDialect"
+    "comm::CommDialect"
   ];
 }
